Add separate directory for interaction arrows
