+++ font/sbolv/interaction-arrows/inhibitorArrow.js
@@ -1,16 +1,16 @@
 
-var Vec2 = require('../../lib/geom/vec2')
-var Rect = require('../../lib/geom/rect')
+var Vec2 = require('../../../lib/geom/vec2')
+var Rect = require('../../../lib/geom/rect')
 
 function createGeometry(boxSize) {
    
     return {
          
-        left: Vec2(boxSize.x,boxSize.y * 0.5),
-        right: Vec2(1.5*boxSize.x, boxSize.y * 0.5),
-        arrowTop:Vec2(1.5 * boxSize.x, boxSize.y * 0.4), 
-        arrowBottom:Vec2(1.5 * boxSize.x, boxSize.y * 0.6),
-        arrowPick:Vec2(boxSize.x * 1.6, boxSize.y * 0.5),
+        top: Vec2(boxSize.x/2,0),
+        bottomLeft: Vec2(boxSize.x/4,(2*boxSize.y)/4),
+        bottomRight: Vec2((boxSize.x * 3 )/4, (2*boxSize.y)/4),
+        middleBottom: Vec2(boxSize.x/2, (2*boxSize.y)/4)
+
     };
 }
 
@@ -20,22 +20,19 @@ function renderGlyph(design, glyphObject, boxSize) {
 
     var path = [
 
-        'M' + Vec2.toPathString(geom.left),
-        'L' + Vec2.toPathString(geom.right),
+        'M' + Vec2.toPathString(geom.top),
+        'L' + Vec2.toPathString(geom.middleBottom),
 
-        'M' + Vec2.toPathString(geom.arrowTop),
-        'L' + Vec2.toPathString(geom.arrowBottom),
-        'L' + Vec2.toPathString(geom.arrowPick),
-        'Z'
+        'M' + Vec2.toPathString(geom.bottomLeft),
+        'L' + Vec2.toPathString(geom.bottomRight)
 
     ].join('');
 
     var glyph = design.surface.path(path);
     var group = design.surface.group()
 
-    glyph.attr('stroke', 'orange');
+    glyph.attr('stroke', 'red');
     glyph.attr('stroke-width', glyphObject.thickness || '3px');
-    glyph.attr('fill', 'none');
     
     group.add(glyph);
