+++ font/sbolv/interaction-arrows/main.js
@@ -1,0 +1,8 @@
+
+module.exports = {
+'InhibitorArrow': require('./inhibitorArrow'),
+'ProductArrow': require('./productArrow'),
+'SimulatorArrow': require('./simulatorArrow'),
+'ReactantArrow': require('./reactantArrow'),
+'DegradationArrow': require('./degradationArrow')
+}
