+++ lib/component.js
@@ -79,8 +79,10 @@ function renderComponent(design, component, interactions) {
 
     entities.forEach(group.add.bind(group));
 
+
     var componenetBoxPadding = 0;
     //design.linkInteractions(group); //has to be commented out , now that interactions are added, it doesn't work
+
     if (interactions.length !== 0 ) {
        componentBoxPadding = 2 * design.geom.componentBoxPadding;
     }
@@ -110,7 +112,7 @@ function renderComponent(design, component, interactions) {
 
     box.radius(8);
 
-    box.attr('stroke', component.color || 'black');
+    box.attr('stroke', component.color || 'white');
     box.attr('stroke-width', component.thickness || '2px');
     box.attr('stroke-linejoin', 'round');
     box.attr('fill', 'none');
