+++ lib/design.js
@@ -25,7 +25,7 @@ function Design(opts) {
         processBoxSize: Vec2.multiply(Vec2(cellSize), 3),
         glyphPadding: cellSize * 4,
         labelOffset: cellSize * 1,
-        componentBoxPadding: cellSize * 1,
+        componentBoxPadding: cellSize * 3,
         arcDistance: cellSize * 1,
         arcEndLength: cellSize * 4,
         arcTurnLength: cellSize * 2,
@@ -141,9 +141,11 @@ Design.prototype = {
         });
 
         var group = this.surface.group();
+        var interactionGroup = this.surface.group();
 
         this.components.forEach(group.add.bind(group));
         this.processes.forEach(group.add.bind(group));
+        this.interactions.forEach(interactionGroup.add.bind(interactionGroup));
 
         //this.linkInteractions(group);
 
@@ -159,7 +161,7 @@ Design.prototype = {
 
         bbox = Rect(group.bbox());
 
-        this.size = Vec2(Rect.width(bbox) + 32, Rect.height(bbox) + 32);
+        this.size = Vec2(Math.max(Rect.width(bbox) + 32, interactionGroup.bbox().width + 32), Rect.height(bbox) + 32);
 
         this.element.style.width = this.size.x + 'px';
         this.element.style.height = this.size.y + 'px';
