+++ lib/displayList.js
@@ -88,14 +88,12 @@ function DisplayList(displayListObject) {
           else if (interaction.type === "non-covelant-binding") {
              
               for (let i in interaction.participants) {
-                 console.log("interaction.participants[i]",interaction.participants[i])
-                
+                                 
                 //find the DNA part in the interaction
                 for ( let j in displayList.components[0].segments ) {
 
                    if (displayList.components[0].segments[j].name === interaction.participants[i].name ) {
                       interaction.participants[i].segment = displayList.components[0].segments[j];
-                      console.log("interaction.participants[i]",interaction.participants[i])
                    }
                } 
              }
