+++ lib/interaction.js
@@ -4,7 +4,7 @@ var Segment = require('./segment'),
     Rect = require('./geom/rect'),
     Vec2 = require('./geom/vec2')
     Matrix = require('./geom/matrix'),
-    Linker = require('./linker'),
+    Linker = require('./linker')
 
    
 function renderInteraction(design, interaction) {
@@ -227,11 +227,8 @@ function renderInteraction(design, interaction) {
     box.attr('fill', 'none');
 
     group.add(box);
-
-    //if (interactionIndex !== 0 ) {
-      //design.geom.interactionPosition = design.geom.interactionPosition - group.bbox().height - design.geom.interactionPadding;
-      design.geom.interactionPosition = design.geom.interactionPosition + group.bbox().height + design.geom.interactionPadding;
-    //}
+    design.geom.interactionPosition = design.geom.interactionPosition + group.bbox().height + design.geom.interactionPadding;
+     
     
     //var interactionLabel = interaction.name || interaction.displayId
     if(interaction.displayId !== undefined) {
