+++ lib/segment.js
@@ -5,11 +5,7 @@ var Rect = require('./geom/rect'),
     ComplexInfo = require('./complexInformation'),
     Composite = require('../font/sbolv/composite'),
     nonDNATypes = require('./nonDNATypes'),
-    Inhibitor = require('../font/sbolv/inhibitor'),
-    Product = require('../font/sbolv/product'),
-    Simulator = require('../font/sbolv/simulator'),
-    ReactantArrow = require('../font/sbolv/reactantArrow'),
-    DegradationArrow = require('../font/sbolv/degradationArrow')
+    Arrows = require('../font/sbolv/interaction-arrows/main')
 
 function renderSegment(design, segment, interactions) {
 
@@ -212,7 +208,7 @@ function renderSegment(design, segment, interactions) {
         //if glyph object has the role of an inhibitor in an interaction, Add inhibitor arrow to the object
         if (segment.role === "inhibitor") {
 
-           var inhibitorProps = Inhibitor.render(design, glyphObject, boundingBoxSize );
+           var inhibitorProps = Arrows.InhibitorArrow.render(design, glyphObject, boundingBoxSize );
            var inhibitorGlyph = inhibitorProps.glyph;
           group.add(inhibitorGlyph);
         }
@@ -220,7 +216,7 @@ function renderSegment(design, segment, interactions) {
         //if glyph object has the role of an template in a production, Add production arrow to the object
         if (segment.role === "product") {
 
-           var productProps = Product.render(design, glyphObject, boundingBoxSize );
+           var productProps = Arrows.ProductArrow.render(design, glyphObject, boundingBoxSize );
            var productGlyph = productProps.glyph;
           group.add(productGlyph);
         }
@@ -228,7 +224,7 @@ function renderSegment(design, segment, interactions) {
         //if glyph object has the role of a simulator in a simulation, Add simulation arrow to the object
         if (segment.role === "simulator") {
 
-           var simulatorProps = Simulator.render(design, glyphObject, boundingBoxSize );
+           var simulatorProps = Arrows.SimulatorArrow.render(design, glyphObject, boundingBoxSize );
            var simulatorGlyph = simulatorProps.glyph;
            group.add(simulatorGlyph);
         }
@@ -238,7 +234,7 @@ function renderSegment(design, segment, interactions) {
            var arrowPos = Matrix();
            var arrowOffset = Vec2(0, -boundingBoxSize.y);
            var arrowTransform = Matrix.multiply( Matrix.translation(arrowOffset), arrowPos );
-           var arrowProps = ReactantArrow.render(design, glyphObject, boundingBoxSize );
+           var arrowProps = Arrows.ReactantArrow.render(design, glyphObject, boundingBoxSize );
            var arrowGlyph = arrowProps.glyph;
            arrowGlyph.transform({ matrix: Matrix.toSVGString(arrowTransform) });
            group.add(arrowGlyph);
@@ -250,7 +246,7 @@ function renderSegment(design, segment, interactions) {
            var arrowPos = Matrix();
            var arrowOffset = Vec2(0, -boundingBoxSize.y);
            var arrowTransform = Matrix.rotate(arrowPos, 180, { x: 20, y: 0 });
-           var arrowProps = ReactantArrow.render(design, glyphObject, boundingBoxSize );
+           var arrowProps = Arrows.ReactantArrow.render(design, glyphObject, boundingBoxSize );
            var arrowGlyph = arrowProps.glyph;
            arrowGlyph.transform({ matrix: Matrix.toSVGString(arrowTransform) });
            group.add(arrowGlyph);
@@ -265,7 +261,7 @@ function renderSegment(design, segment, interactions) {
            var arrowPos = Matrix();
            var arrowOffset = Vec2(0, -boundingBoxSize.y);
            var arrowTransform = Matrix.multiply( Matrix.translation(arrowOffset), arrowPos );
-           var arrowProps = DegradationArrow.render(design, glyphObject, boundingBoxSize );
+           var arrowProps = Arrows.DegradationArrow.render(design, glyphObject, boundingBoxSize );
            var arrowGlyph = arrowProps.glyph;
            arrowGlyph.transform({ matrix: Matrix.toSVGString(arrowTransform) });
            group.add(arrowGlyph);
